refactor(pdf-export): extract dialog state factories from PdfExportDialog and PdfPreviewDialog
Moves all reactive state, derived values, effects, and functions out of
PdfExportDialog.svelte (317-line script → 43 lines) and PdfPreviewDialog.svelte
(61-line script → 14 lines) into co-located state factory files following the
create{Name}State() pattern. Also removes local interface re-definitions in
PdfPreviewDialog in favour of the shared types/components/PdfPreviewDialog.d.ts,
and uses PreviewState type in PdfExportDialogState instead of the inline duplicate.
